feat: support --force option when archiving projects
