+++ frontend/src/locales/en-US.json
@@ -1455,7 +1455,10 @@
       "archive": {
         "title": "Archive project",
         "description": "Archived project will not be shown on the normal interface. You can click \"Archive\" on the bottom left to find and restore it.",
-        "btn-text": "Archive this @.lower:{'common.project'}"
+        "btn-text": "Archive this @.lower:{'common.project'}",
+        "force": {
+          "description": "Force to archive. All releated issues will be resolved, and all related sheets will be deleted."
+        }
       },
       "restore": {
         "title": "Restore project",
